LPD-57535 Extracting test to a separate class
